Zip generator is now an async background process called through curl

+++ includes/filebox.php
@@ -180,6 +180,8 @@ Login and change you settings to unsubscribe from these emails.', 'filebox' )
 		add_action( 'wp_ajax_filebox_reset_file', array( &$this, 'reset_file' ) );
 		// Reset file
 		add_action( 'wp_ajax_filebox_delete_file', array( &$this, 'delete_file' ) );
+		// Generate zip
+		add_action( 'wp_ajax_filebox_generate_zip', array( &$this, 'generate_folder_zip' ) );
 
 		/**
 		 * Folder actions
@@ -1225,19 +1227,80 @@ Login and change you settings to unsubscribe from these emails.', 'filebox' )
 	 * @param int $folder_id
 	 * @return void
 	 */
-	public function generate_folder_zip( $folder_id ) {
-		$zip = new ZipArchive();
-		$filename = $this->get_folder_zip_path( $folder_id );
+	public function generate_folder_zip( $args ) {
+		set_time_limit( 0 );
+		$now = time();
 
-		if( is_file( $filename ) ) {
-			unlink( $filename );
+		if( is_numeric( $args ) ) {
+			$folder_id = $args;
+		} else {
+			$args = $this->get_ajax_arguments( $args, array(
+				'folder_id' => 0
+			), 'generate_folder_zip' );
+
+			$folder_id = $args[ 'folder_id' ];
 		}
 
-		if( $zip->open( $filename, ZIPARCHIVE::CREATE ) ) {
-			$this->generate_folder_zip_files( $zip, $folder_id );
+		if( $folder_id ) {
+			$zip = new ZipArchive();
+			$filename = preg_replace(
+				'/\.zip$/',
+				sprintf( '.%s.zip', $now ),
+				$this->get_folder_zip_path( $folder_id )
+			);
+
+			if( is_file( $filename ) ) {
+				unlink( $filename );
+			}
+
+			if( $zip->open( $filename, ZIPARCHIVE::CREATE ) ) {
+				$this->generate_folder_zip_files( $zip, $folder_id );
+			}
+
+			$zip->close();
+			$target = $this->get_folder_zip_path( $folder_id );
+
+			if( file_exists( $target ) ) {
+				unlink( $target );
+			}
+
+			symlink( $filename, $target );
+
+			foreach( glob( preg_match(
+				'/\.zip$/', '.*.zip', $this->get_folder_zip_path( $folder_id )
+			) ) as $symlink ) {
+				if( preg_match( '/\.([0-9]+)\.zip$/', $symlink, $time ) ) {
+					$time = ( int ) $time[ 0 ];
+
+					if( $time - 86400 < $now && $symlink != $filename ) {
+						unlink( $symlink );
+					}
+				}
+			}
+		}
+	}
+
+	/**
+	 * Starts an async call to the generate_folder_zip via an ajax request.
+	 * @param int $folder_id
+	 * @return void
+	 */
+	public function generate_folder_zip_async( $folder_id ) {
+		global $ajaxurl;
+		$nonce = wp_create_nonce( 'generate_folder_zip' );
+		$cookies = array();
+
+		foreach( $_COOKIE as $key => $val ) {
+			$cookies[] = sprintf( '%s=%s', $key, urlencode( $val ) );
 		}
 
-		$zip->close();
+		exec( sprintf(
+			'curl -b "%s" -d "action=filebox_generate_zip" -d "security=%s" -d "folder_id=%s" "%s" > /dev/null 2>&1 &',
+			implode( ';', $cookies ),
+			$nonce,
+			$folder_id,
+			admin_url( 'admin-ajax.php' )
+		) );
 	}
 
 	/**
@@ -1423,7 +1486,7 @@ Login and change you settings to unsubscribe from these emails.', 'filebox' )
 
 			if( $args[ 'folder_id' ] ) {
 				if( ! is_file( $this->get_folder_zip_path( $args[ 'folder_id' ] ) ) ) {
-					$this->generate_folder_zip( $args[ 'folder_id' ] );
+					$this->generate_folder_zip_async( $args[ 'folder_id' ] );
 				}
 
 				$response[ 'folders' ] = $this->get_subfolders( $args[ 'folder_id' ] );
@@ -1568,7 +1631,7 @@ Login and change you settings to unsubscribe from these emails.', 'filebox' )
 				$group_id = $this->get_group_by_folder( $args[ 'folder_id' ] );
 				$group_folder_id = $this->get_group_folder( $group_id );
 
-				$this->generate_folder_zip( $group_folder_id );
+				$this->generate_folder_zip_async( $group_folder_id );
 
 				do_action(
 					'filebox_file_upload',
@@ -1627,7 +1690,7 @@ Login and change you settings to unsubscribe from these emails.', 'filebox' )
 					$group_id = $this->get_group_by_folder( $args[ 'folder_id' ] );
 					$group_folder_id = $this->get_group_folder( $group_id );
 
-					$this->generate_folder_zip( $group_folder_id );
+					$this->generate_folder_zip_async( $group_folder_id );
 
 					do_action(
 						'filebox_move_file',
@@ -1689,7 +1752,7 @@ Login and change you settings to unsubscribe from these emails.', 'filebox' )
 				$group_id = $this->get_group_by_file( $args[ 'file_id' ] );
 				$group_folder_id = $this->get_group_folder( $group_id );
 
-				$this->generate_folder_zip( $group_folder_id );
+				$this->generate_folder_zip_async( $group_folder_id );
 
 				do_action(
 					'filebox_rename_file',
@@ -1810,7 +1873,7 @@ Login and change you settings to unsubscribe from these emails.', 'filebox' )
 				$group_id = $this->get_group_by_file( $args[ 'file_id' ] );
 				$group_folder_id = $this->get_group_folder( $group_id );
 
-				$this->generate_folder_zip( $group_folder_id );
+				$this->generate_folder_zip_async( $group_folder_id );
 			} else {
 				$response[ 'error' ] = __( 'Error when trying to trash file.', 'filebox' );
 			}
@@ -1970,7 +2033,7 @@ Login and change you settings to unsubscribe from these emails.', 'filebox' )
 				$group_id = $this->get_group_by_folder( $args[ 'folder_id' ] );
 				$group_folder_id = $this->get_group_folder( $group_id );
 
-				$this->generate_folder_zip( $group_folder_id );
+				$this->generate_folder_zip_async( $group_folder_id );
 			} else {
 				$response[ 'error' ] = __( 'Error when trying to move folder.', 'filebox' );
 			}
@@ -2059,7 +2122,7 @@ Login and change you settings to unsubscribe from these emails.', 'filebox' )
 
 		if( wp_delete_term( $args[ 'folder_id' ], 'fileboxfolders' ) ) {
 			$response[ 'folder_id' ] = $args[ 'folder_id' ];
-			$this->generate_folder_zip( $group_folder_id );
+			$this->generate_folder_zip_async( $group_folder_id );
 		} else {
 			$response[ 'error' ] = __( 'Error when trying to delete folder.', 'filebox' );
 		}
